+++ lib/samanage/api.rb
@@ -109,7 +109,13 @@ module Samanage
 			# Error cases
 			case response[:code]
 			when 200..201
-				response[:data] = JSON.parse(api_call.body)
+				begin
+					response[:data] = JSON.parse(api_call.body)
+				rescue JSON::ParserError => e
+					response[:data] = api_call.body
+					puts "** Warning **#{e.class}"
+					puts e
+				end
 				response
 			when 401
 				response[:data] = api_call.body